requirements.txt
asgiref==3.7.2
beautifulsoup4==4.12.2
cffi==1.15.1
cryptography==40.0.2
Django==3.2.20
django-bootstrap3==23.2
django-bootstrap4==23.1
django-crispy-forms==2.0
gitdb==4.0.10
GitPython==3.1.31
importlib-metadata==2.1.3
mysql-connector-python==8.0.33
Pillow==9.5.0
protobuf==3.20.3
pycparser==2.21
PyMySQL==1.0.3
pytz==2023.3
smmap==5.0.0
soupsieve==2.4.1
sqlparse==0.4.4
typing_extensions==4.6.3
zipp==3.15.0
